# docs/swagger.json
{
    "schemes": [
        "https",
        "http"
    ],
    "swagger": "2.0",
    "info": {
        "description": "Dedicated API for ArcheryHub Mobile App (Archer \u0026 Scorekeeper)",
        "title": "ArcheryHub Mobile API",
        "termsOfService": "http://archeryhub.id/terms/",
        "contact": {
            "name": "ArcheryHub Support",
            "url": "https://archeryhub.id",
            "email": "support@archeryhub.id"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.1"
    },
    "host": "api.archeryhub.id",
    "basePath": "/api/v1",
    "paths": {
...
        "/mobile/auth/archer/login": {
            "post": {
                "description": "Login using email and password for archers",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Auth"
                ],
                "summary": "Archer Login",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.mobileEmailPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/auth/archer/register": {
            "post": {
                "description": "Register a new archer account",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Auth"
                ],
                "summary": "Archer Registration",
                "parameters": [
                    {
                        "description": "Registration Details",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileArcherRegisterRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileLoginResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/auth/google/login": {
            "post": {
                "description": "Authenticate or register using Google ID Token",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Auth"
                ],
                "summary": "Google Login",
                "parameters": [
                    {
                        "description": "Google ID Token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileOAuthLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/auth/organization/login": {
            "post": {
                "description": "Login for organization accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Auth"
                ],
                "summary": "Organization Login",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.mobileEmailPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/auth/scorekeeper/login": {
            "post": {
                "description": "Login using a numeric code assigned by organization",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Auth"
                ],
                "summary": "Scorekeeper Login",
                "parameters": [
                    {
                        "description": "Scorekeeper Code",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileScorekeeperLoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/mobile/auth/seller/login": {
            "post": {
                "description": "Login for seller accounts",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Mobile - Auth"
                ],
                "summary": "Seller Login",
                "parameters": [
                    {
                        "description": "Login Credentials",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/mobile.mobileEmailPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/mobile.MobileLoginResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "mobile.MobileArcherRegisterRequest": {
            "type": "object",
            "properties": {
                "avatar_url": {
                    "type": "string",
                    "example": "https://cdn.archeryhub.id/media/archers/rizky.jpg"
                },
                "birth_date": {
                    "type": "string",
                    "example": "1995-05-18"
                },
                "email": {
                    "type": "string",
                    "example": "archer@example.com"
                },
                "full_name": {
                    "type": "string",
                    "example": "Rizky Pratama"
                },
                "gender": {
                    "type": "string",
                    "example": "male"
                },
                "password": {
                    "type": "string",
                    "example": "securepassword123"
                },
                "phone": {
                    "type": "string",
                    "example": "081234567890"
                },
                "username": {
                    "type": "string",
                    "example": "rizky-pratama"
                }
            }
        },
...
        "mobile.MobileLoginResponse": {
            "type": "object",
            "properties": {
                "is_new_user": {
                    "type": "boolean"
                },
                "token": {
                    "type": "string"
                },
                "user": {
                    "$ref": "#/definitions/mobile.MobileUser"
                }
            }
        },
...
        "mobile.MobileOAuthLoginRequest": {
            "type": "object",
            "required": [
                "idToken"
            ],
            "properties": {
                "idToken": {
                    "type": "string"
                }
            }
        },
...
        "mobile.mobileEmailPasswordRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "ApiKeyAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}